refactor(typechecker): split storage, scope, alias, and registry helpers
Move inferred-type storage, scope/symbol helpers, aliasing/display helpers,
and Defs registry helpers out of typechecker.go into inferred_storage.go,
scope_symbol.go, type_alias_display.go, and type_registry.go so the core
checker file stays focused on CheckTypes and construction.

feat(typechecker): infer shape literals with callee parameter context

When argument count matches the callee signature, infer call arguments with
the formal parameter type so struct literals pick up fields like *String for
AppMutation.Input-shaped parameters. After inferring a function body, sync
tc.Functions[name].Parameters with resolved parameter types. Prefer expected
named types in inferShapeType when they match structurally; use
shapesHaveSameStructure / IsTypeCompatible so hash-named shapes align with
named aliases and non-scalar values can satisfy *T (scalars still need explicit
pointers).

Example:

```forst
func createTask(op AppMutation.Input({
  ctx: { sessionId: &sessionId, user: { name: "Alice" } },
  input: { name: "Fix leak" },
})) { /* ... */ }
```

fix(typechecker): Result+ensure inference and Value(Ref) pointer fields

Count ensure statements via stmt.Kind() so *ast.EnsureNode is included. When
ensure composes with a plain success return, infer a single Result(S, Error)
instead of a legacy (S, error) pair; storeInferredFunctionReturnType keeps the
declared Result when the body infers plain S. For Value constraints that
reference variables, infer Pointer(T) for the field type so &sessionId-shaped
literals stay pointer-backed.

Example:

```forst
func okInt() {
  n := 42
  ensure n is GreaterThan(0)
  return n
}
```

test(examples): exercise ensure lowering in Result examples and refresh guard golden

Add ensure n is GreaterThan(0) to result_if / result_ensure helpers; add
TestResultExamplesIncludeEnsureLowering and TestCheckTypes_ensureWithoutExplicitReturn_infersSingleResult.
Simplify shape_guard.ft call sites and update examples/out/rfc/guard/shape_guard.go.

chore(vscode): add Kill Forst LSP task

Add a shell task that pkill's forst lsp for local LSP restarts; minor launch.json whitespace.
