+++ .vscode/launch.json
@@ -1,6 +1,7 @@
 {
   "version": "0.2.0",
   "configurations": [
+    
     {
       "name": "Run Extension (vscode-forst)",
       "type": "extensionHost",
