+++ .vscode/tasks.json
@@ -2,6 +2,16 @@
   "version": "2.0.0",
   "tasks": [
     {
+      "label": "Kill Forst LSP",
+      "type": "shell",
+      "command": "pkill -f 'forst lsp' 2>/dev/null || true; echo 'Forst LSP processes stopped (or none were running).'",
+      "problemMatcher": [],
+      "presentation": {
+        "reveal": "silent",
+        "panel": "shared"
+      }
+    },
+    {
       "label": "compile: vscode-forst",
       "type": "shell",
       "command": "bun run --filter forst compile",
