+++ examples/in/result_if.ft
@@ -2,8 +2,11 @@ package main
 
 // Result split + if x is Ok() narrowing; golden: ../out/result_if/result_if.go
 
-func one(): Result(Int, Error) {
-	return 1
+func one() {
+	// Ensure in the body keeps Result/ensure tooling and narrowing aligned; return type is inferred as Result(Int, Error).
+	n := 1
+	ensure n is GreaterThan(0)
+	return n
 }
 
 func main() {
