+++ forst/internal/typechecker/go_builtins.go
@@ -438,6 +438,20 @@ func (tc *TypeChecker) IsTypeCompatible(actual ast.TypeNode, expected ast.TypeNo
 		return true
 	}
 
+	// Value compatible with *T when compatible with T for non-scalar T (shape literals for *User,
+	// etc.). Scalars still require an explicit pointer so String does not satisfy *String.
+	if expected.Ident == ast.TypePointer && len(expected.TypeParams) == 1 {
+		inner := expected.TypeParams[0]
+		if !isScalarTypeIdent(inner.Ident) && tc.IsTypeCompatible(actual, inner) {
+			tc.log.WithFields(logrus.Fields{
+				"actual":   actual.Ident,
+				"expected": expected.Ident,
+				"function": "IsTypeCompatible",
+			}).Debug("Actual type compatible with pointer element type")
+			return true
+		}
+	}
+
 	// Check if actual type is an alias of expected type
 	actualDef, actualExists := tc.Defs[actual.Ident]
 	if actualExists {
@@ -498,7 +512,9 @@ func (tc *TypeChecker) IsTypeCompatible(actual ast.TypeNode, expected ast.TypeNo
 		}).Info("Shape extraction results")
 
 		if actualShapeOk && expectedShapeOk {
-			identical := tc.shapesAreStructurallyIdentical(*actualShape, *expectedShape)
+			// Prefer shapesHaveSameStructure: it resolves assertion fields and uses assignability for
+			// mismatched type identifiers (e.g. inferred literal ctx vs AppContext).
+			identical := tc.shapesHaveSameStructure(*actualShape, *expectedShape)
 			tc.log.WithFields(logrus.Fields{
 				"actual":    actual.Ident,
 				"expected":  expected.Ident,
@@ -541,6 +557,15 @@ func (tc *TypeChecker) IsTypeCompatible(actual ast.TypeNode, expected ast.TypeNo
 	return false
 }
 
+func isScalarTypeIdent(id ast.TypeIdent) bool {
+	switch id {
+	case ast.TypeString, ast.TypeInt, ast.TypeFloat, ast.TypeBool:
+		return true
+	default:
+		return false
+	}
+}
+
 // getShapeFromTypeDef extracts the shape from a TypeDefNode if it's a shape definition
 func (tc *TypeChecker) getShapeFromTypeDef(def ast.Node) (*ast.ShapeNode, bool) {
 	if typeDef, ok := def.(ast.TypeDefNode); ok {
@@ -568,9 +593,15 @@ func (tc *TypeChecker) shapesAreStructurallyIdentical(a, b ast.ShapeNode) bool {
 				// Unknown types are compatible with any concrete type
 				continue
 			}
-			if fieldA.Type.Ident != fieldB.Type.Ident {
-				return false
+			if fieldA.Type.Ident == fieldB.Type.Ident {
+				continue
 			}
+			// Hash-based structural types vs named shapes (e.g. inferred literal vs AppContext) must use
+			// full assignability, not identifier equality only.
+			if tc.IsTypeCompatible(*fieldA.Type, *fieldB.Type) {
+				continue
+			}
+			return false
 		} else if fieldA.Shape != nil && fieldB.Shape != nil {
 			if !tc.shapesAreStructurallyIdentical(*fieldA.Shape, *fieldB.Shape) {
 				return false
